Moved TargetCreate, TargetViewAll, and TargetView to their own directories. Also converted CSS to SCSS.
